Create Forecast Weather ViewModel and Fragment

+++ app/src/main/java/com/example/umbrella/model/currentweatherresponse/Wind.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Wind(
 
 	@field:SerializedName("deg")
-	val deg: Int? = null,
+	val deg: Double? = null,
 
 	@field:SerializedName("speed")
 	val speed: Double? = null
